Update Player.cs
Added player gravity + ground constant into code. The ground is currently hardcoded, and we'll establish and replace it later with a global constant in game. The jump currently operates with GetKeyHeld, not GetKeyUp. This is because it just looks a lot smoother with held, but if we find a better group of constants to replace it, we'll replace GetKeyHeld with GetKeyUp.